requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
fastjsonschema==2.21.1
filelock==3.20.0
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
inflection==0.5.1
ipykernel==6.30.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==1.0.0
langchain-upstage==0.7.4
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
lark==1.2.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
matplotlib-inline==0.2.1
mistune==3.1.3
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.3.5
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
pillow==11.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
selenium==4.35.0
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.44
sqlparse==0.5.4
stack-data==0.6.3
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.20.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.14
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
